requirements.txt
appnope==0.1.0
autopep8==1.2.2
bcrypt==2.0.0
blinker==1.4
cffi==1.5.2
coverage==4.0.3
decorator==4.0.9
dominate==2.1.16
Flask==0.10.1
Flask-Admin==1.4.0
Flask-Bootstrap==3.3.5.7
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
gnureadline==6.3.3
gunicorn==19.4.5
honcho==0.6.6
ipdb==0.8.3
ipython==4.1.1
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
marshmallow==2.6.0
mccabe==0.4.0
passlib==1.6.5
path.py==8.1.2
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
py==1.4.31
pycparser==2.14
pyflakes==1.0.0
pytest==2.8.7
pytest-cov==2.2.1
pytest-flask==0.10.0
python-dateutil==2.4.2
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.11
SQLAlchemy-Utils==0.31.6
traitlets==4.1.0
visitor==0.1.2
Werkzeug==0.11.3
wheel==0.24.0
WTForms==2.1
